# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/users/login" : {
      "post" : {
        "summary" : "登录",
        "description" : "登录",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "captcha",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/register" : {
      "post" : {
        "summary" : "注册",
        "description" : "注册",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "username" : {
                    "type" : "string",
                    "description" : "用户账号"
                  },
                  "password" : {
                    "type" : "string",
                    "description" : "密码"
                  },
                  "role" : {
                    "type" : "string",
                    "description" : "用户类型"
                  },
                  "addtime" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/updatePassword" : {
      "get" : {
        "summary" : "修改密码",
        "description" : "修改密码",
        "parameters" : [ {
          "name" : "oldPassword",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "newPassword",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/resetPass" : {
      "get" : {
        "summary" : "密码重置",
        "description" : "密码重置",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/page" : {
      "get" : {
        "summary" : "列表",
        "description" : "列表",
        "parameters" : [ {
          "name" : "key.key",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "",
            "nullable" : false
          }
        }, {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "username",
          "in" : "query",
          "description" : "用户账号",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "密码",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "role",
          "in" : "query",
          "description" : "用户类型",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "addtime",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/list" : {
      "get" : {
        "summary" : "列表",
        "description" : "列表",
        "parameters" : [ {
          "name" : "id",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "nullable" : false
          }
        }, {
          "name" : "username",
          "in" : "query",
          "description" : "用户账号",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "password",
          "in" : "query",
          "description" : "密码",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "role",
          "in" : "query",
          "description" : "用户类型",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        }, {
          "name" : "addtime",
          "in" : "query",
          "description" : "",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/info/{id}" : {
      "get" : {
        "summary" : "信息",
        "description" : "信息",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/save" : {
      "post" : {
        "summary" : "保存",
        "description" : "保存",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "username" : {
                    "type" : "string",
                    "description" : "用户账号"
                  },
                  "password" : {
                    "type" : "string",
                    "description" : "密码"
                  },
                  "role" : {
                    "type" : "string",
                    "description" : "用户类型"
                  },
                  "addtime" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/update" : {
      "post" : {
        "summary" : "修改",
        "description" : "修改",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "id" : {
                    "type" : "integer",
                    "description" : ""
                  },
                  "username" : {
                    "type" : "string",
                    "description" : "用户账号"
                  },
                  "password" : {
                    "type" : "string",
                    "description" : "密码"
                  },
                  "role" : {
                    "type" : "string",
                    "description" : "用户类型"
                  },
                  "addtime" : {
                    "type" : "string",
                    "description" : ""
                  }
                },
                "description" : ""
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/delete" : {
      "post" : {
        "summary" : "删除",
        "description" : "删除",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "description" : "",
                "items" : {
                  "type" : "integer"
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "key" : {
                      "type" : "object",
                      "properties" : { }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}